% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPDetect.R
\name{GPDetect}
\alias{GPDetect}
\title{Peak detection using cubic smoothing splines}
\usage{
GPDetect(
  x,
  SNP = SNP,
  Position = Position,
  Chromosome = Chromosome,
  Value = Value,
  savePlots = FALSE,
  res_dir = "results",
  cv = FALSE,
  ...
)
}
\arguments{
\item{x}{a data frame containing single-SNP based GWAS results.}

\item{SNP}{a string denoting the column name for the SNP ids.}

\item{Position}{a string denoting the column name for map position of the SNPs. Said column must be numeric.}

\item{Chromosome}{a string denoting the column name for the chromosome.}

\item{Value}{a string denoting the column name for the test statistic values. Said column must be numeric.
\strong{Value} may refer to any test-statistic value.}

\item{savePlots}{if TRUE, a scatter plot of raw test statistic values is provided along with a smoothed curve fitted over those points. The plot is saved into the specified directory \strong{res_dir}.}

\item{res_dir}{the directory where plots will be saved if desired.}

\item{cv}{if TRUE, the ordinary leave-one-out and if FALSE, the generalized cross-validation is used for smoothing parameter computation.}

\item{...}{further parameters for the \strong{smooth.spline} function of base R.}
}
\value{
A data frame containing the peaks found in the data with their corresponding parameters.
For every peak a start and an end position, ids of the SNPs corresponding to those positions and the total number of SNPs within the peak are given along with the height value.
}
\description{
The function fits cubic smoothing splines on the given test statictic values to detect the patterns inherent in these values.
The inflection points of the curve are then used to define peaks.
}
\details{
The function implements cubic smoothing splines on the test statistic values to obtain a smoothed curve.
The inflection points of the curve are determined and the region between two consecutive inflection points having a downward concave form is regarded as a peak.
The maximum test statistic value within a peak after smoothing is recorded as the height of the peak.
To determine the optimal level of smoothing, cross validation is implemented.
The GWAS results must be obtained beforehand and supplied to the function.
Any test statictic that represents the strength of association between a SNP and the phenotype can be used.
}
\examples{
#retrieve sample file from package
file <- system.file('extdata', 'sim_02D3.assoc', package = 'GPDetect')
df <- read.table(file, sep = ' ', header = TRUE)

#run GPDetect on the data frame df and save plots to the 'results' directory
result <- GPDetect(df, SNP, Position, Chromosome, WaldStat, savePlots = TRUE)
result

}
